requirements.txt
accelerate==0.20.3
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
h2==4.2.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==1.25.0
nvidia-ml-py==13.590.44
openai==2.14.0
packaging==25.0
pandas==1.5.2
peft==0.3.0
Pillow==10.0.1
pip==25.3
platformdirs==4.4.0
prettytable==3.16.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pyecharts==2.0.9
pyg-lib==0.2.0+pt113cu117
Pygments==2.19.2
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0
regex==2025.11.3
requests==2.32.5
rich==13.9.4
safetensors==0.7.0
scikit-learn==1.3.0
scipy==1.13.1
sentence-transformers==5.1.2
sentencepiece==0.2.1
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
swanlab==0.7.6
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.3.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.46.0
typing_extensions==4.14.1
typing-inspection==0.4.2
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
wrapt==2.0.1
yarl==1.22.0
zstandard==0.23.0
